pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "slidegeist"
version = "2025.11.4"
description = "Extract slides and transcripts from lecture videos using external STT and VLM services"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Christopher Albert" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "numpy>=2.2.0,<3.0",
    "opencv-python>=4.10.0,<5.0",
    "yt-dlp>=2024.12.0",
    "pytesseract>=0.3.13,<1.0",
    "Pillow>=11.0.0,<12.0",
    "tqdm>=4.67.0,<5.0",
    "psutil>=7.0.0,<8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
slidegeist = "slidegeist.cli:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=slidegeist --cov-report=term-missing -m 'not manual'"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "manual: marks tests requiring manual intervention or validation",
]
